Rename `#[pin_project]` to `#[pin_v2]` to avoid naming conflicts
