+++ compiler/rustc_attr_parsing/src/attributes/pin_v2.rs
@@ -7,15 +7,15 @@ use crate::context::Stage;
 use crate::target_checking::AllowedTargets;
 use crate::target_checking::Policy::Allow;
 
-pub(crate) struct PinProjectParser;
+pub(crate) struct PinV2Parser;
 
-impl<S: Stage> NoArgsAttributeParser<S> for PinProjectParser {
-    const PATH: &[Symbol] = &[sym::pin_project];
+impl<S: Stage> NoArgsAttributeParser<S> for PinV2Parser {
+    const PATH: &[Symbol] = &[sym::pin_v2];
     const ON_DUPLICATE: OnDuplicate<S> = OnDuplicate::Warn;
     const ALLOWED_TARGETS: AllowedTargets = AllowedTargets::AllowList(&[
         Allow(Target::Enum),
         Allow(Target::Struct),
         Allow(Target::Union),
     ]);
-    const CREATE: fn(Span) -> AttributeKind = AttributeKind::PinProject;
+    const CREATE: fn(Span) -> AttributeKind = AttributeKind::PinV2;
 }
