+++ compiler/rustc_hir_typeck/messages.ftl
@@ -227,9 +227,9 @@ hir_typeck_pass_to_variadic_function = can't pass `{$ty}` to variadic function
     .suggestion = cast the value to `{$cast_ty}`
     .teach_help = certain types, like `{$ty}`, must be cast before passing them to a variadic function to match the implicit cast that a C compiler would perform as part of C's numeric promotion rules
 
-hir_typeck_project_on_non_pin_project_type = cannot project on type that is not `#[pin_project]`
+hir_typeck_project_on_non_pin_project_type = cannot project on type that is not `#[pin_v2]`
     .note = type defined here
-    .suggestion = add `#[pin_project]` here
+    .suggestion = add `#[pin_v2]` here
 
 hir_typeck_ptr_cast_add_auto_to_object = cannot add {$traits_len ->
     [1] auto trait {$traits}
