+++ compiler/rustc_span/src/symbol.rs
@@ -1667,7 +1667,7 @@ symbols! {
         pin,
         pin_ergonomics,
         pin_macro,
-        pin_project,
+        pin_v2,
         platform_intrinsics,
         plugin,
         plugin_registrar,
