+++ tests/ui/pin-ergonomics/pattern-matching-mix-deref-pattern.rs
@@ -10,16 +10,16 @@
 
 use std::pin::Pin;
 
-#[cfg_attr(any(pin_ergonomics, both), pin_project)]
+#[cfg_attr(any(pin_ergonomics, both), pin_v2)]
 struct Foo<T, U> {
     x: T,
     y: U,
 }
 
-#[cfg_attr(any(pin_ergonomics, both), pin_project)]
+#[cfg_attr(any(pin_ergonomics, both), pin_v2)]
 struct Bar<T, U>(T, U);
 
-#[cfg_attr(any(pin_ergonomics, both), pin_project)]
+#[cfg_attr(any(pin_ergonomics, both), pin_v2)]
 enum Baz<T, U> {
     Foo(T, U),
     Bar { x: T, y: U },
@@ -120,27 +120,27 @@ fn bar_const<T: Unpin, U: Unpin>(bar: Pin<&Bar<T, U>>) {
 fn non_pin_project<T, U: Unpin>(foo: Pin<&mut NonPinProject<T>>, bar: Pin<&NonPinProject<U>>) {
     let NonPinProject { x } = foo;
     //[normal]~^ ERROR mismatched types
-    //[pin_ergonomics]~^^ ERROR cannot project on type that is not `#[pin_project]`
-    //[both]~^^^ ERROR cannot project on type that is not `#[pin_project]`
+    //[pin_ergonomics]~^^ ERROR cannot project on type that is not `#[pin_v2]`
+    //[both]~^^^ ERROR cannot project on type that is not `#[pin_v2]`
     let NonPinProject { x } = bar;
     //[normal]~^ ERROR mismatched types
-    //[pin_ergonomics]~^^ ERROR cannot project on type that is not `#[pin_project]`
-    //[both]~^^^ ERROR cannot project on type that is not `#[pin_project]`
+    //[pin_ergonomics]~^^ ERROR cannot project on type that is not `#[pin_v2]`
+    //[both]~^^^ ERROR cannot project on type that is not `#[pin_v2]`
 
     match foo {
         NonPinProject { x } => {}
         //[normal]~^ ERROR mismatched types
         //[deref_patterns]~^^ ERROR mix of deref patterns and normal constructors
-        //[pin_ergonomics]~^^^ ERROR cannot project on type that is not `#[pin_project]`
-        //[both]~^^^^ ERROR cannot project on type that is not `#[pin_project]`
+        //[pin_ergonomics]~^^^ ERROR cannot project on type that is not `#[pin_v2]`
+        //[both]~^^^^ ERROR cannot project on type that is not `#[pin_v2]`
         Pin { .. } => {}
     }
     match bar {
         NonPinProject { x } => {}
         //[normal]~^ ERROR mismatched types
         //[deref_patterns]~^^ ERROR mix of deref patterns and normal constructors
-        //[pin_ergonomics]~^^^ ERROR cannot project on type that is not `#[pin_project]`
-        //[both]~^^^^ ERROR cannot project on type that is not `#[pin_project]`
+        //[pin_ergonomics]~^^^ ERROR cannot project on type that is not `#[pin_v2]`
+        //[both]~^^^^ ERROR cannot project on type that is not `#[pin_v2]`
         Pin { .. } => {}
     }
 }
