+++ tests/ui/pin-ergonomics/pattern-matching.normal.stderr
@@ -1,18 +1,18 @@
-error[E0658]: the `#[pin_project]` attribute is an experimental feature
+error[E0658]: the `#[pin_v2]` attribute is an experimental feature
   --> $DIR/pattern-matching.rs:12:1
    |
-LL | #[pin_project]
-   | ^^^^^^^^^^^^^^
+LL | #[pin_v2]
+   | ^^^^^^^^^
    |
    = note: see issue #130494 <https://github.com/rust-lang/rust/issues/130494> for more information
    = help: add `#![feature(pin_ergonomics)]` to the crate attributes to enable
    = note: this compiler was built on YYYY-MM-DD; consider upgrading it if it is out of date
 
-error[E0658]: the `#[pin_project]` attribute is an experimental feature
+error[E0658]: the `#[pin_v2]` attribute is an experimental feature
   --> $DIR/pattern-matching.rs:18:1
    |
-LL | #[pin_project]
-   | ^^^^^^^^^^^^^^
+LL | #[pin_v2]
+   | ^^^^^^^^^
    |
    = note: see issue #130494 <https://github.com/rust-lang/rust/issues/130494> for more information
    = help: add `#![feature(pin_ergonomics)]` to the crate attributes to enable
